Half-fixed feedback net drawing
Needed to decouple net vs edge name when building the edges that
connect the dummy nodes introduced for long edges.

Edge now has a 'name' field. This is usually the same as its old 'net'
attribute, but when adding an edge to connect dummy vertices, the 'name'
changes. And from then on in the program, Edges are keyed by .name, not .net