Sort've fixed feedback node drawing.
Still can't do crossover minimisation properly when an edge
joins to the same layer. So it can't figure out that an inner
loop is possible in loopback_3.

Need to make sure all the dummy edges are resized to the maximum
or mimimum width of modules in its layer to fix up other drawing
problems that I've introduced in these changes.